+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/Option.java
@@ -302,7 +302,14 @@ public enum Option {
      *
      * @since 4.6.0
      */
-    ALLOF_CLEANUP_AT_THE_END(null, null);
+    ALLOF_CLEANUP_AT_THE_END(null, null),
+    /**
+     * Whether at the end of the schema generation, all sub-schemas without an explicit "type" indication should be augmented by the implied "type"
+     * based on the other tags in the respective schema.
+     *
+     * @since 4.30.0
+     */
+    STRICT_TYPE_INFO(null, null);
 
     /**
      * Optional: the module realising the setting/option if it is enabled.